alarm/antilles-mail-agent/requirements.txt
falcon>=1.3
PyYAML>=3.10
setuptools>=36.0
jsonschema>=2.5.1